Refactor for gjenbruk av copyContentNode
